gametest/requirements.txt
wheel
Django
django-rest-framework
psycopg2-binary
